pyproject.toml
[project]
name = "quickpde"
description = ""
readme = "README.md"
requires-python = ">=3.11"
version = "0.1.0"

dependencies = [
  "numpy",
  "jax",
  "h5py",
  "matplotlib",
  "hydra-core",
  "hydra-submitit-launcher",
  "humanize",
  "pandas",
  "seaborn",
  "imageio[ffmpeg] (>=2.37.2,<3.0.0)",
  "tqdm",
  "sphinx-autobuild (>=2025.8.25,<2026.0.0)",
  "zarr (>=3.1.5,<4.0.0)",
]

[project.optional-dependencies]
dev = [
  "pylint>=2.6.0",
  "yapf",
  "pytest",
  "isort",
  "sphinx",
  "furo",
  "myst-parser",
  "sphinx-autodoc-typehints",
  "sphinx-copybutton",
  "sphinx-mathjax-offline",
  "sphinx-autobuild"
]

[project.scripts]
analysis = "quickpde.cli:main"

[tool.setuptools]
packages = ["quickpde"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
